[cxs] SPLIT HEADER mgOnGpuTypes.h into mgOnGpuFptypes.h and mgOnGpuCxtypes.h (but keep a wrapper of both)
