+++ epochX/cudacpp/ee_mumu/src/mgOnGpuFptypes.h
@@ -1,0 +1,72 @@
+#ifndef MGONGPUFPTYPES_H
+#define MGONGPUFPTYPES_H 1
+
+#include "mgOnGpuConfig.h"
+
+#include <cmath>
+
+// --- Functions and operators for floating point types
+
+#ifdef __CUDACC__ // cuda
+
+/*
+inline __host__ __device__
+fptype fpmax( const fptype& a, const fptype& b )
+{
+  return max( a, b );
+}
+
+inline __host__ __device__
+fptype fpmin( const fptype& a, const fptype& b )
+{
+  return min( a, b );
+}
+*/
+
+inline __host__ __device__
+const fptype& fpmax( const fptype& a, const fptype& b )
+{
+  return ( ( b < a ) ? a : b );
+}
+
+inline __host__ __device__
+const fptype& fpmin( const fptype& a, const fptype& b )
+{
+  return ( ( a < b ) ? a : b );
+}
+
+inline __host__ __device__
+fptype fpsqrt( const fptype& f )
+{
+#if defined MGONGPU_FPTYPE_FLOAT
+  // See https://docs.nvidia.com/cuda/cuda-math-api/group__CUDA__MATH__SINGLE.html
+  return sqrtf( f );
+#else
+  // See https://docs.nvidia.com/cuda/cuda-math-api/group__CUDA__MATH__DOUBLE.html
+  return sqrt( f );
+#endif
+}
+
+#else // c++
+
+inline
+const fptype& fpmax( const fptype& a, const fptype& b )
+{
+  return std::max( a, b );
+}
+
+inline
+const fptype& fpmin( const fptype& a, const fptype& b )
+{
+  return std::min( a, b );
+}
+
+inline
+fptype fpsqrt( const fptype& f )
+{
+  return std::sqrt( f );
+}
+
+#endif
+
+#endif // MGONGPUFPTYPES_H
